requirements.txt
# Core Dependencies for the project
django 
djangorestframework 
django-rest-knox
django-redis
whitenoise       
django-cors-headers
dj-database-url

# Environment and configuration
python-dotenv
python-json-logger

# Date and time handling
pytz              

# External services integration
geopy              
requests           

# Testing dependencies 
pytest       
pytest-django      
pytest-cov
requests-mock      
mock     

# Utility
packaging
